more accurate deps parsing for FreeBSD git installs

+++ bootstrap-salt.sh
@@ -5698,8 +5698,11 @@ install_freebsd_git_deps() {
 
     if [ "${_POST_NEON_INSTALL}" -eq $BS_FALSE ]; then
 
-        SALT_DEPENDENCIES=$(/usr/local/sbin/pkg search -R -d py37-salt | grep 'origin:' \
-            | tail -n +2 | awk -F\" '{print $2}')
+        if ! __check_command_exists jq; then
+            /usr/local/sbin/pkg install -y jq || return 1
+        fi
+
+        SALT_DEPENDENCIES=$(/usr/local/sbin/pkg search -R -d --raw-format json py37-salt | jq -r '.deps|keys[]')
         # shellcheck disable=SC2086
         /usr/local/sbin/pkg install -y ${SALT_DEPENDENCIES} python || return 1
